Check config directory exists before copying prod.env.js file

+++ install.js
@@ -19,4 +19,6 @@ if (!fs.existsSync(env_example)) {
 }
 
 copySync('env.example.js', path.join(base_dir, 'env.js'));
-copySync('config.prod.env.js', path.join(base_dir, 'config', 'prod.env.js'), true);
+if (fs.existsSync(path.join(base_dir, 'config'))) {
+    copySync('config.prod.env.js', path.join(base_dir, 'config', 'prod.env.js'), true);
+}
